# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  db.CaasWithdrawal:
    properties:
      amount:
        type: string
      caas_withdrawal_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      idempotency_key:
        type: string
      payout_mobile:
        type: string
      payout_network:
        type: string
      phone:
        type: string
      status:
        type: string
      token:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  handlers.BalanceData:
    properties:
      balance_usdc:
        example: "100.000000"
        type: string
      wallet_address:
        example: 0xSCW123...
        type: string
    type: object
...
  handlers.CaasWalletData:
    properties:
      abstraction_address:
        example: 0xSCW123...
        type: string
      caas_wallet_id:
        example: 0xblindindex...
        type: string
      created_at:
        type: string
      id:
        type: string
      is_active:
        example: true
        type: boolean
      user_id:
        type: string
    type: object
  handlers.CaasWalletResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.CaasWalletData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.CryptoBalanceResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.BalanceData'
      success:
        example: true
        type: boolean
    type: object
  handlers.CryptoTxData:
    properties:
      amount:
        example: "50.00"
        type: string
      created_at:
        type: string
      id:
        type: string
      receiver_phone:
        example: "+237698765432"
        type: string
      reference:
        example: CRYPTO-REF-001
        type: string
      sender_user_id:
        type: string
      status:
        example: pending
        type: string
      token:
        example: USDC
        type: string
      tx_hash:
        type: string
    type: object
  handlers.CryptoTxListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.CryptoTxData'
        type: array
      success:
        example: true
        type: boolean
    type: object
  handlers.CryptoTxResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.CryptoTxData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        properties:
          code:
            example: VALIDATION_ERROR
            type: string
          message:
            example: phone is required
            type: string
        type: object
      success:
        example: false
        type: boolean
    type: object
...
  handlers.FundAccountRequest:
    properties:
      amount:
        description: Amount in local fiat (minimum 100)
        example: 20000
        minimum: 100
        type: number
      currency:
        description: 'Currency of the Mobile Money deposit: XOF or XAF'
        enum:
        - XOF
        - XAF
        example: XOF
        type: string
      operator:
        description: Operator is the Mobile Money provider
        enum:
        - Orange
        - MTN
        - Wave
        - Moov
        - Airtel
        example: MTN
        type: string
      phone:
        description: Phone is the Mobile Money number to pull from (E.164)
        example: "+22507000000"
        type: string
      token:
        description: Token is the target stablecoin — defaults to USDC
        enum:
        - USDC
        - USDT
        example: USDC
        type: string
    type: object
...
  handlers.Hub2RefData:
    properties:
      hub2_reference:
        example: HUB2-REF-001
        type: string
    type: object
  handlers.Hub2RefResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.Hub2RefData'
      success:
        example: true
        type: boolean
    type: object
...
  handlers.SendCryptoRequest:
    properties:
      amount:
        example: "50.00"
        type: string
      receiver_phone:
        example: "+237698765432"
        type: string
      token:
        description: Token must be USDT or USDC
        enum:
        - USDT
        - USDC
        example: USDC
        type: string
    type: object
...
  handlers.WithdrawCryptoRequest:
    properties:
      amount:
        description: Amount is the stablecoin amount to withdraw (decimal string)
        example: "50.00"
        type: string
      payout_mobile:
        description: PayoutMobile is the Mobile Money number to disburse fiat to (E.164)
        example: "+22507000000"
        type: string
      payout_network:
        description: PayoutNetwork is the MNO operator
        enum:
        - Orange
        - MTN
        - Wave
        - Moov
        - Airtel
        example: MTN
        type: string
      token:
        description: Token must be USDT or USDC
        enum:
        - USDC
        - USDT
        example: USDC
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: engineering@rach.finance
    name: Rach Finance Engineering
  description: |-
    DigitalFX is a multi-currency fintech platform for Central/West Africa.
    It combines fiat accounts (XAF, XOF, USD, GBP, EUR), HD crypto wallets (BIP-44 via Rach WaaS),
    and stablecoin P2P transfers (USDC/USDT via Rach CaaS ERC-4337 SCWs).
    Mobile Money deposits and withdrawals are handled via HUB2 (MTN, Orange).
  title: DigitalFX API
  version: "1.0"
paths:
...
  /crypto/balances:
    get:
      description: Returns the caller's Instant USD (iUSD) balance from Rach CaaS.
        iUSD settles on-chain as USDC inside the EIP-4337 Smart Contract Wallet, but
        is always shown to the customer as iUSD. This is separate from any USDT/USDC
        held in the Rach WaaS crypto wallets.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CryptoBalanceResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get iUSD balance
      tags:
      - CaaS - Instant USD (iUSD)
...
  /crypto/fund:
    post:
      consumes:
      - application/json
      description: 'Funds the customer''s Instant USD (iUSD) account. Initiates a
        HUB2 Mobile Money collection: the customer receives a push-to-pay prompt on
        their phone. After they approve, HUB2 fires a webhook which triggers Rach
        CaaS to convert the XOF/XAF to on-chain USDC and credit the customer''s EIP-4337
        Smart Contract Wallet — reflected to the customer as iUSD. Poll GET /crypto/balances
        for the updated iUSD balance once CaaS confirms the fiat and completes the
        conversion.'
      parameters:
      - description: Funding request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.FundAccountRequest'
      produces:
      - application/json
      responses:
        "201":
          description: HUB2 collection reference — poll status or wait for balance
            update
          schema:
            $ref: '#/definitions/handlers.Hub2RefResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Fund iUSD account via Mobile Money
      tags:
      - CaaS - Instant USD (iUSD)
  /crypto/send:
    post:
      consumes:
      - application/json
      description: 'Transfers Instant USD (iUSD) from the caller to another DigitalFX
        user identified by phone number, via Rach CaaS (settles on-chain as USDC between
        the two EIP-4337 wallets). Amount is USD-equivalent decimal (e.g. "50.00").
        Note: the `token` field selects the on-chain settlement asset (USDC/USDT)
        but the customer-facing unit is iUSD. This is the CaaS P2P rail, distinct
        from WaaS on-chain crypto transfers.'
      parameters:
      - description: Transfer details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.SendCryptoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.CryptoTxResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send iUSD to another user (Phone Send)
      tags:
      - CaaS - Instant USD (iUSD)
  /crypto/transactions:
    get:
      description: Returns a paginated list of the user's Instant USD (iUSD / CaaS)
        transactions — funding, Phone Send transfers, and off-ramp withdrawals. These
        are CaaS-rail movements, not WaaS on-chain crypto history (see /wallets/crypto/{network}/transactions
        for that).
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page, max 100 (default 20)
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CryptoTxListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List iUSD (CaaS) transactions
      tags:
      - CaaS - Instant USD (iUSD)
  /crypto/transactions/{id}:
    get:
      description: Returns a single Instant USD (iUSD / CaaS) transaction by its UUID.
      parameters:
      - description: Transaction UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CryptoTxResponse'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get an iUSD (CaaS) transaction
      tags:
      - CaaS - Instant USD (iUSD)
  /crypto/wallet:
    get:
      description: Returns the caller's Instant USD (iUSD) wallet — an EIP-4337 Smart
        Contract Wallet (SCW) provisioned on Rach CaaS, keyed by the user's phone
        number (provisioned on first call). This is the CaaS rail (instant dollars
        that settle on-chain as USDC), NOT the Rach WaaS HD crypto wallets. Customer-facing
        unit is always iUSD.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CaasWalletResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get or create the user's iUSD (CaaS) wallet
      tags:
      - CaaS - Instant USD (iUSD)
  /crypto/withdraw:
    post:
      consumes:
      - application/json
      description: Off-ramps Instant USD (iUSD) from the user's EIP-4337 Smart Contract
        Wallet back to fiat on their Mobile Money number, via Rach CaaS (settles the
        on-chain USDC and pays out local currency). The `token` field is the on-chain
        settlement asset; the customer balance is iUSD.
      parameters:
      - description: Withdrawal details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.WithdrawCryptoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.CaasWithdrawal'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Withdraw iUSD to Mobile Money (off-ramp)
      tags:
      - CaaS - Instant USD (iUSD)
...
securityDefinitions:
  BearerAuth:
    description: 'JWT access token — prefix with "Bearer ": Bearer eyJhbGci...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"